Move the CPU friction laws into their own folder
