+++ src/DynamicRupture/FrictionLaws/CpuImpl/FastVelocityWeakeningLaw.h
