+++ src/DynamicRupture/FrictionLaws/CpuImpl/NoFault.cpp
