+++ src/DynamicRupture/FrictionLaws/CpuImpl/NoFault.h
