+++ src/DynamicRupture/FrictionLaws/CpuImpl/SourceTimeFunction.cpp
