+++ src/DynamicRupture/FrictionLaws/CpuImpl/SourceTimeFunction.h
