+++ src/DynamicRupture/FrictionLaws/CpuImpl/ThermalPressurization/NoTP.h
